pyproject.toml
[tool.poetry]
name = "testbot"
version = "0.1.0"
description = ""
authors = ["John Peng"]
readme = "README.md"

packages = [
    { include = "testbot", from = "src" }
]

[tool.poetry.scripts]
testbot = "testbot.cli:main"

[tool.poetry.dependencies]
python = "^3.12"
mypy = "^1.14.1"
sqlmodel = "^0.0.22"
tiktoken = "^0.8.0"
litellm = "^1.59.3"
pydantic = "^2.10.5"
gitpython = "^3.1.44"
braintrust = "^0.0.183"
colorama = "^0.4.6"
pytz = "^2024.2"
instructor = {path = "../instructor", develop = true}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"